Smaller commit to clean up use case 3, add tests, and make sure it adheres to checkstyle. hopefully this is the part 2 (and final part) of use case 3.
