+++ src/main/java/gui/Sakura.java
@@ -1,0 +1,16 @@
+package gui;
+
+import java.awt.Color;
+
+/** Sakura theme.
+ *  Pastel pink colors
+ */
+class Sakura extends Palette {
+    public Sakura() {
+        this.mainBg = new Color(253, 247, 250);
+        this.panelBg = new Color(246, 230, 236);
+        this.text = new Color(74, 58, 65);
+        this.buttonBg = new Color(235, 167, 192);
+        this.buttonText = new Color(255, 255, 255);
+    }
+}
